Flatten Header Layout
Having the wmde logo alone on its row costs us
expensive vertical screen estate and the save
button may be below the fold.

Thus the logo is reintegrated into the root
component. At some point we should extract this
as a new header component.

This commit also adds the question mark,
because with a third item we get nicer styling
with flexbox.